third_party/WebKit/LayoutTests/http/tests/bluetooth/https/requestDevice/cross-origin-iframe.html

<!DOCTYPE html>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script src="/js-test-resources/testdriver.js"></script>
<script src="/js-test-resources/testdriver-vendor.js"></script>
<script src="/bluetooth-resources/bluetooth-helpers.js"></script>
<body>
<script>
'use strict';
const test_desc = 'Request device from a unique origin. ' +
  'Should reject with SecurityError.';
let iframe = document.createElement('iframe');

bluetooth_test(() => setUpConnectableHealthThermometerDevice()
    // 1. Load the iframe.
    .then(() => new Promise(resolve => {
      iframe.src = 'https://localhost:8443/js-test-resources/bluetooth/' +
        'requestDevice-in-iframe.html';
      document.body.appendChild(iframe);
      iframe.addEventListener('load', resolve);
    }))
    // 2. Request the device from the iframe.
    .then(() => new Promise(resolve => {
      callWithTrustedClick(() => {
        iframe.contentWindow.postMessage('Go', '*');
      });

      window.onmessage = messageEvent => {
        assert_equals(messageEvent.data, 'SecurityError: requestDevice() ' +
            'called from cross-origin iframe.');
        resolve();
      }
    })), test_desc);
</script>
</body>
